Group related routes

+++ api/v1/views/amenities.py
@@ -6,7 +6,6 @@ from flask import jsonify, request
 from models import storage
 from models.amenity import Amenity
 
-
 @app_views.route('/amenities', strict_slashes=False)
 def amenities():
     """Returns a JSON string containing all amenity objects
@@ -16,27 +15,32 @@ def amenities():
     return jsonify(amenities)
 
 
-@app_views.route('/amenities/<amenity_id>', strict_slashes=False)
-def amenity(amenity_id):
-    """Returns a JSON string containing an amenity object
-    """
-    amenity = storage.get("Amenity", amenity_id)
-    if amenity is None:
-        return jsonify({"error": "Not found"}), 404
-    return jsonify(amenity.to_dict())
-
-
 @app_views.route('/amenities/<amenity_id>', strict_slashes=False,
-                 methods=['DELETE'])
-def delete_amenity(amenity_id):
-    """Deletes an amenity object
+                 methods=['GET', 'DELETE', 'PUT'])
+def amenity(amenity_id):
+    """Handles GET and DELETE and PUT requests for an amenity object
     """
     amenity = storage.get("Amenity", amenity_id)
     if amenity is None:
         return jsonify({"error": "Not found"}), 404
-    storage.delete(amenity)
+    if request.method == 'GET':
+        return jsonify(amenity.to_dict())
+    
+    if request.method == 'DELETE':
+        storage.delete(amenity)
+        storage.save()
+        return jsonify({}), 200
+    
+    if request.method == 'PUT':
+        data = request.get_json(silent=True)
+    if data is None:
+        return jsonify({"error": "Not a JSON"}), 400
+    
+    for key, value in data.items():
+        if key not in ["id", "created_at", "updated_at"]:
+            setattr(amenity, key, value)
     storage.save()
-    return jsonify({}), 200
+    return jsonify(amenity.to_dict()), 200
 
 
 @app_views.route('/amenities', strict_slashes=False, methods=['POST'])
@@ -52,22 +56,3 @@ def create_amenity():
     amenity = Amenity(**data)
     amenity.save()
     return jsonify(amenity.to_dict()), 201
-
-
-@app_views.route('/amenities/<amenity_id>', strict_slashes=False,
-                 methods=['PUT'])
-def update_amenity(amenity_id):
-    """Updates an amenity object
-    """
-    amenity = storage.get("Amenity", amenity_id)
-    if amenity is None:
-        return jsonify({"error": "Not found"}), 404
-    data = request.get_json(silent=True)
-    if data is None:
-        return jsonify({"error": "Not a JSON"}), 400
-
-    for key, value in data.items():
-        if key not in ["id", "created_at", "updated_at"]:
-            setattr(amenity, key, value)
-    storage.save()
-    return jsonify(amenity.to_dict()), 200
