+++ api/v1/views/cities.py
@@ -18,33 +18,38 @@ def cities(state_id):
     return jsonify(cities)
 
 
-@app_views.route('/cities/<city_id>', strict_slashes=False)
-def city(city_id):
-    """Returns a JSON string containing a city object
-    """
-    city = storage.get("City", city_id)
-    if city is None:
-        return jsonify({"error": "Not found"}), 404
-    return jsonify(city.to_dict())
-
-
 @app_views.route('/cities/<city_id>', strict_slashes=False,
-                 methods=['DELETE'])
-def delete_city(city_id):
-    """Deletes a city object
+                 methods=['GET', 'DELETE', 'PUT'])
+def city(city_id):
+    """Handles GET, DELETE and PUT requests for a city object
     """
     city = storage.get("City", city_id)
     if city is None:
         return jsonify({"error": "Not found"}), 404
-    storage.delete(city)
+    if request.method == 'GET':
+        return jsonify(city.to_dict())
+    
+    if request.method == 'DELETE':
+        storage.delete(city)
+        storage.save()
+        return jsonify({}), 200
+    
+    if request.method == 'PUT':
+        data = request.get_json(silent=True)
+    if data is None:
+        return jsonify({"error": "Not a JSON"}), 400
+    
+    for key, value in data.items():
+        if key not in ["id", "created_at", "updated_at"]:
+            setattr(city, key, value)
     storage.save()
-    return jsonify({}), 200
+    return jsonify(city.to_dict()), 200
 
 
 @app_views.route('/states/<state_id>/cities', strict_slashes=False,
                  methods=['POST'])
 def create_city(state_id):
-    """Creates a city object
+    """Creates a city object, linked to a state
     """
     state = storage.get("State", state_id)
     if state is None:
@@ -59,22 +64,3 @@ def create_city(state_id):
     city.state_id = state_id
     city.save()
     return jsonify(city.to_dict()), 201
-
-
-@app_views.route('/cities/<city_id>', strict_slashes=False,
-                 methods=['PUT'])
-def update_city(city_id):
-    """Updates a city object
-    """
-    city = storage.get("City", city_id)
-    if city is None:
-        return jsonify({"error": "Not found"}), 404
-    data = request.get_json(silent=True)
-    if data is None:
-        return jsonify({"error": "Not a JSON"}), 400
-
-    for key, value in data.items():
-        if key not in ["id", "created_at", "updated_at"]:
-            setattr(city, key, value)
-    storage.save()
-    return jsonify(city.to_dict()), 200
